Updated module dependency-order algorithm.
The previous algorithm guaranteed that the modules list of a firmware
would be in dependency order.  The new algorithm makes it more likely that
low level modules will be more toward the front of the list, when there
are higher level modules that don't depend on that lower level module.

This is not a bug fix since the previous order was correct.  The new order
appears to be more correct intuitively and less likely to cause confusion.